Refactor interactive package
- Moved ca and cert prompts to their own sub-packages
